+++ src/app/public/modules/alert/alert.component.ts
@@ -2,8 +2,21 @@ import {
   Component,
   EventEmitter,
   Input,
+  OnInit,
   Output
 } from '@angular/core';
+
+import {
+  SkyIndicatorIconType
+} from '../shared/indicator-icon-type';
+
+import {
+  SkyIndicatorIconUtility
+} from '../shared/indicator-icon-utility';
+
+import {
+  SkyIconStackItem
+} from '../icon/icon-stack-item';
 
 const ALERT_TYPE_DEFAULT = 'warning';
 
@@ -12,7 +25,7 @@ const ALERT_TYPE_DEFAULT = 'warning';
   styleUrls: ['./alert.component.scss'],
   templateUrl: './alert.component.html'
 })
-export class SkyAlertComponent {
+export class SkyAlertComponent implements OnInit {
   /**
    * Specifies a style for the alert to determine the icon and background color.
    * The valid options are `danger`, `info`, `success`, and `warning`.
@@ -21,6 +34,7 @@ export class SkyAlertComponent {
   @Input()
   public set alertType(value: string) {
     this._alertType = value;
+    this.updateAlertIcon();
   }
 
   public get alertType(): string {
@@ -47,10 +61,27 @@ export class SkyAlertComponent {
   @Output()
   public closedChange = new EventEmitter<boolean>();
 
+  public alertBaseIcon: SkyIconStackItem;
+
+  public alertTopIcon: SkyIconStackItem;
+
   private _alertType: string;
+
+  public ngOnInit(): void {
+    this.updateAlertIcon();
+  }
 
   public close(): void {
     this.closed = true;
     this.closedChange.emit(true);
+  }
+
+  private updateAlertIcon(): void {
+    const indicatorIcon = SkyIndicatorIconUtility.getIconsForType(
+      this.alertType as SkyIndicatorIconType
+    );
+
+    this.alertBaseIcon = indicatorIcon.modernThemeBaseIcon;
+    this.alertTopIcon = indicatorIcon.modernThemeTopIcon;
   }
 }