+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 4.2.0 (2020-06-19)
+
+- Added the status indicator component to replace the previous CSS version. [#91](https://github.com/blackbaud/skyux-indicators/pull/91)
+
+# 4.1.0 (2020-06-09)
+
+- Added modern theme styles to the alert component. [#87](https://github.com/blackbaud/skyux-indicators/pull/87)
+
 # 4.0.0 (2020-05-12)
 
 ### New features